Legge om og erstatte intl og FormattedMessage med typede komponenter
