+++ apps/sif-ettersending/src/app/i18n/index.tsx
@@ -1,0 +1,52 @@
+import { FormattedMessage, useIntl } from 'react-intl';
+import { commonMessages } from '@navikt/sif-common-core-ds/src/i18n/common.messages';
+import { typedIntlHelper } from '@navikt/sif-common-core-ds/src/utils/intlUtils';
+import { soknadMessages } from '@navikt/sif-common-soknad-ds';
+import { uiMessages } from '@navikt/sif-common-ui';
+import { appMessages } from './appMessages';
+
+export const libMessages = {
+    nb: {
+        ...commonMessages.nb,
+        ...uiMessages.nb,
+        ...soknadMessages.nb,
+    },
+    nn: {
+        ...commonMessages.nn,
+        ...uiMessages.nn,
+        ...soknadMessages.nn,
+    },
+};
+
+const nb = {
+    ...libMessages.nb,
+    ...appMessages.nb,
+};
+
+const nn: Record<keyof typeof nb, string> = {
+    ...libMessages.nn,
+    ...appMessages.nn,
+};
+
+export type AppMessageKeys = keyof typeof nb;
+
+export const useAppIntl = () => {
+    const intl = useIntl();
+    return typedIntlHelper<AppMessageKeys>(intl);
+};
+
+export type AppIntlShape = ReturnType<typeof useAppIntl>;
+
+interface AppTextProps {
+    id: AppMessageKeys;
+    values?: any;
+}
+
+export const AppText = (props: AppTextProps) => {
+    return <FormattedMessage {...props} />;
+};
+
+export const applicationIntlMessages = {
+    nb,
+    nn,
+};
