+++ apps/sif-ettersending/src/app/pages/velkommen-page/OmSøknaden.tsx
@@ -1,24 +1,36 @@
 import { BodyLong, Heading, Link } from '@navikt/ds-react';
-import { FormattedMessage } from 'react-intl';
+import { ReactNode } from 'react';
 import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
+import { AppText } from '../../i18n';
 import getLenker from '../../lenker';
 
 const OmSøknaden = () => {
     return (
         <Block margin="xl">
             <Heading level="2" size="medium">
-                <FormattedMessage id="page.velkommen.omSøknaden.tittel" />
+                <AppText id="page.velkommen.omSøknaden.tittel" />
             </Heading>
             <BodyLong as="div">
-                <FormattedMessage id="page.velkommen.omSøknaden.1" tagName="p" />
-                <FormattedMessage id="page.velkommen.omSøknaden.2" tagName="p" />
-                <FormattedMessage id="page.velkommen.omSøknaden.3" tagName="p" />
+                <p>
+                    <AppText id="page.velkommen.omSøknaden.1" />
+                </p>
+                <p>
+                    <AppText id="page.velkommen.omSøknaden.2" />
+                </p>
+                <p>
+                    <AppText id="page.velkommen.omSøknaden.3" />
+                </p>
                 <Block>
-                    <FormattedMessage id="page.velkommen.omSøknaden.4.1" />
-                    <Link href={getLenker().personvern} target="_blank">
-                        <FormattedMessage id="page.velkommen.omSøknaden.4.2" />
-                    </Link>
-                    <FormattedMessage id="page.velkommen.omSøknaden.4.3" />
+                    <AppText
+                        id="page.velkommen.omSøknaden.4"
+                        values={{
+                            Lenke: (children: ReactNode) => (
+                                <Link href={getLenker().personvern} target="_blank">
+                                    {children}
+                                </Link>
+                            ),
+                        }}
+                    />
                 </Block>
             </BodyLong>
         </Block>
