+++ apps/sif-ettersending/src/app/pages/velkommen-page/VelkommenGuide.tsx
@@ -1,15 +1,15 @@
-import { Ingress, GuidePanel, Heading } from '@navikt/ds-react';
+import { BodyLong, GuidePanel, Heading } from '@navikt/ds-react';
 import React from 'react';
-import { FormattedMessage } from 'react-intl';
+import { AppText } from '../../i18n';
 
 const VelkommenGuide: React.FunctionComponent = () => (
     <GuidePanel>
         <Heading level="1" size="large" spacing={true}>
-            <FormattedMessage id="page.velkommen.guide.tittel" />
+            <AppText id="page.velkommen.guide.tittel" />
         </Heading>
-        <Ingress>
-            <FormattedMessage id="page.velkommen.guide.ingress" />
-        </Ingress>
+        <BodyLong size="large">
+            <AppText id="page.velkommen.guide.ingress" />
+        </BodyLong>
     </GuidePanel>
 );
