--- apps/sif-ettersending/src/app/pages/velkommen-page/velkommenPageMessages.ts
@@ -1,22 +1,0 @@
-import { MessageFileFormat } from '@navikt/sif-common-core-ds/src/types/MessageFileFormat';
-
-export const velkommenPageMessages: MessageFileFormat = {
-    nb: {
-        'page.velkommen.guide.tittel': 'Hei',
-        'page.velkommen.guide.ingress':
-            'Her laster du opp dokumenter som hører til en søknad du har sendt inn tidligere. Har du dokumentene klare?',
-
-        'page.velkommen.omSøknaden.tittel': 'Om ettersendingen',
-        'page.velkommen.omSøknaden.1': 'Du får veiledning underveis om hva du skal fylle ut, og hvordan.',
-        'page.velkommen.omSøknaden.fremoverITid':
-            'Hvis du søker frem i tid, har vi forståelse for at du kan bli usikker på hva du skal svare på noen spørsmål som handler om tiden fremover. Da svarer du så godt du kan ut fra hva situasjonen er nå, og så kan du melde fra om eventuelle endringer senere.',
-        'page.velkommen.omSøknaden.2':
-            'Vi tar vare på svarene dine i 72 timer. Hvis du innenfor den tiden for eksempel vil ta en pause eller blir automatisk logget ut, fortsetter du der du var når du kommer tilbake.',
-        'page.velkommen.omSøknaden.3': 'Du må svare på alle spørsmålene for å kunne gå videre.',
-
-        'page.velkommen.omSøknaden.4.1':
-            'Du har rett til innsyn i saken din. Vil du vite mer om hvordan NAV behandler personopplysninger? Se ',
-        'page.velkommen.omSøknaden.4.2': 'nav.no/personvern',
-        'page.velkommen.omSøknaden.4.3': '.',
-    },
-};
