+++ apps/sif-ettersending/src/app/soknad/beskrivelse-step/BeskrivelseStep.tsx
@@ -1,9 +1,8 @@
 import { BodyLong, Link } from '@navikt/ds-react';
 import React from 'react';
-import { FormattedMessage, useIntl } from 'react-intl';
 import FormBlock from '@navikt/sif-common-core-ds/src/atoms/form-block/FormBlock';
-import intlHelper from '@navikt/sif-common-core-ds/src/utils/intlUtils';
 import { getStringValidator } from '@navikt/sif-common-formik-ds/src/validation';
+import { AppText, useAppIntl } from '../../i18n';
 import { SoknadFormField } from '../../types/SoknadFormData';
 import { Søknadstype } from '../../types/Søknadstype';
 import { MAX_BESKRIVELSE_LENGTH, MIN_BESKRIVELSE_LENGTH } from '../../validation/fieldValidations';
@@ -16,14 +15,14 @@ interface Props {
 }
 
 const BeskrivelseStep: React.FC<Props> = ({ søknadstype }) => {
-    const intl = useIntl();
+    const { text } = useAppIntl();
     return (
         <SoknadFormStep id={StepID.BESKRIVELSE} søknadstype={søknadstype}>
             <FormBlock>
                 <SoknadFormComponents.Textarea
                     data-testid="beskrivelse"
                     name={SoknadFormField.beskrivelse}
-                    label={intlHelper(intl, 'step.beskrivelse.hvaSendes.spm')}
+                    label={text('step.beskrivelse.hvaSendes.spm')}
                     maxLength={MAX_BESKRIVELSE_LENGTH}
                     autoComplete="off"
                     validate={(value) => {
@@ -45,18 +44,18 @@ const BeskrivelseStep: React.FC<Props> = ({ søknadstype }) => {
                     description={
                         <BodyLong as="div">
                             <p>
-                                <FormattedMessage id="step.beskrivelse.intro.1" />
+                                <AppText id="step.beskrivelse.intro.1" />
                             </p>
                             <ul>
                                 <li>
-                                    <FormattedMessage id="step.beskrivelse.intro.li.1" />
+                                    <AppText id="step.beskrivelse.intro.li.1" />
                                 </li>
                                 <li>
-                                    <FormattedMessage id="step.beskrivelse.intro.li.2" />
+                                    <AppText id="step.beskrivelse.intro.li.2" />
                                 </li>
                             </ul>
                             <p>
-                                <FormattedMessage
+                                <AppText
                                     id="step.beskrivelse.intro.2"
                                     values={{
                                         kontaktMedOssLink: (
@@ -64,7 +63,7 @@ const BeskrivelseStep: React.FC<Props> = ({ søknadstype }) => {
                                                 target="_blank"
                                                 rel="noopener noreferrer"
                                                 href={'https://www.nav.no/person/kontakt-oss/nb/'}>
-                                                <FormattedMessage id="step.beskrivelse.intro.2.1" />
+                                                <AppText id="step.beskrivelse.intro.2.1" />
                                             </Link>
                                         ),
                                     }}
