+++ apps/sif-ettersending/src/app/soknad/beskrivelsePP-step/AnnetBarnPart.tsx
@@ -1,13 +1,11 @@
 import { Heading } from '@navikt/ds-react';
 import React from 'react';
-import { useIntl } from 'react-intl';
 import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
-import intlHelper from '@navikt/sif-common-core-ds/src/utils/intlUtils';
 import { SkjemagruppeQuestion } from '@navikt/sif-common-formik-ds/src';
 import { getFødselsnummerValidator } from '@navikt/sif-common-formik-ds/src/validation';
-
-import SoknadFormComponents from '../SoknadFormComponents';
+import { useAppIntl } from '../../i18n';
 import { SoknadFormField } from '../../types/SoknadFormData';
+import SoknadFormComponents from '../SoknadFormComponents';
 
 interface Props {
     søkersFødselsnummer: string;
@@ -15,7 +13,7 @@ interface Props {
 }
 
 const AnnetBarnPart: React.FC<Props> = ({ søkersFødselsnummer, harRegistrerteBarn }) => {
-    const intl = useIntl();
+    const { text } = useAppIntl();
 
     return (
         <Block margin="xl">
@@ -23,13 +21,13 @@ const AnnetBarnPart: React.FC<Props> = ({ søkersFødselsnummer, harRegistrerteB
                 legend={
                     harRegistrerteBarn ? (
                         <Heading level="2" size="small" style={{ display: 'inline-block', fontSize: '1.125rem' }}>
-                            {intlHelper(intl, 'step.beskrivelse_pp.annetBarn.tittel')}
+                            {text('step.beskrivelse_pp.annetBarn.tittel')}
                         </Heading>
                     ) : undefined
                 }>
                 <div>
                     <SoknadFormComponents.TextField
-                        label={intlHelper(intl, 'step.beskrivelse_pp.annetBarn.fnr.spm')}
+                        label={text('step.beskrivelse_pp.annetBarn.fnr.spm')}
                         name={SoknadFormField.barnetsFødselsnummer}
                         validate={getFødselsnummerValidator({
                             required: true,
