+++ apps/sif-ettersending/src/app/validation/apiValuesValidation.ts
@@ -1,7 +1,6 @@
-import { IntlShape } from 'react-intl';
 import { ValidationSummaryError } from '@navikt/sif-common-core-ds/src/components/validation-error-summary-base/ValidationErrorSummaryBase';
-import intlHelper from '@navikt/sif-common-core-ds/src/utils/intlUtils';
 import { SoknadApiData } from '../types/SoknadApiData';
+import { AppIntlShape } from '../i18n';
 
 export const apiVedleggIsInvalid = (vedlegg: string[]): boolean => {
     vedlegg.find((v) => {
@@ -10,13 +9,16 @@ export const apiVedleggIsInvalid = (vedlegg: string[]): boolean => {
     return vedlegg.length === 0 || vedlegg.find((v) => v === undefined || v === '' || v === null) !== undefined;
 };
 
-export const validateApiValues = (values: SoknadApiData, intl: IntlShape): ValidationSummaryError[] | undefined => {
+export const validateApiValues = (
+    values: SoknadApiData,
+    { text }: AppIntlShape,
+): ValidationSummaryError[] | undefined => {
     const errors: ValidationSummaryError[] = [];
 
     if (apiVedleggIsInvalid(values.vedlegg)) {
         errors.push({
             name: 'vedlegg',
-            message: intlHelper(intl, 'steg.oppsummering.validering.manglerVedlegg'),
+            message: text('steg.oppsummering.validering.manglerVedlegg'),
         });
     }
     return errors.length > 0 ? errors : undefined;
